+++ src/main/java/ru/gdim/simple_java_rest_api/service/resource_version/ResourceVersionService.java
@@ -56,6 +56,6 @@ public final class ResourceVersionService {
             return null;
         }
 
-        return new ResourceVersion(url, eTag, lastModified);
+        return new ResourceVersionImpl(url, eTag, lastModified);
     }
 }
